Create from list of Genomes now
Instead of directly looping through a range of random seeds to create the different display images, the code now loops through a list of Genomes instead. For the moment, the Genomes are created by looping through a list of seeds, but later the Genome list could be initialized differently or mutated, etc.